requirements.txt
streamlit>=1.54.0,<2.0.0
requests>=2.32.5,<3.0.0
python-dotenv>=1.2.2,<2.0.0
pydantic>=2.12.5,<3.0.0
beautifulsoup4>=4.14.3,<5.0.0
chromadb>=1.5.2,<2.0.0
langchain>=1.2.10,<2.0.0
langchain-google-genai>=4.2.1,<5.0.0
langchain-community>=0.4.1,<0.5.0
langchain-text-splitters>=1.1.1,<2.0.0
langgraph>=1.0.10,<2.0.0

